requirements.txt
torch
torchvision
torchaudio
timm
scikit-learn
numpy
pandas
Pillow
opencv-python
albumentations
matplotlib
seaborn
pyyaml
tqdm
streamlit
wandb
tensorboard
tensorflow